y2038: timerfd: Use timespec64 internally
timerfd_show() uses a 'struct itimerspec' internally, but that is
deprecated because of the time_t overflow and a conflict with the glibc
type of the same name that is now incompatible in user space.

Use a pair of timespec64 variables instead as a simple replacement.

As this removes the last use of itimerspec from the kernel, allowing the
removal of the definition from the uapi headers along with timespec and
timeval later.

Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
(cherry picked from commit bde9e963af2d0a9762781c4dccaef21696d5bb94)

Change-Id: I2d09df00a5ea616703e74465ca3627b8c939ffa5
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>