drm/msm: avoid using 'timespec'
The timespec structure and associated interfaces are deprecated and will
be removed in the future because of the y2038 overflow.

The use of ktime_to_timespec() in timeout_to_jiffies() does not
suffer from that overflow, but is easy to avoid by just converting
the ktime_t into jiffies directly.

Reviewed-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
(cherry picked from commit 6cedb8b377bba6520508399b74754b9a0ca06edf)

Change-Id: I1f9067bf4443f2f7cc8e8b83f2159c0e80762a7a
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>