y2038: uapi: change __kernel_time_t to __kernel_old_time_t
This is mainly a patch for clarification, and to let us remove
the time_t definition from the kernel to prevent new users from
creeping in that might not be y2038-safe.

All remaining uses of 'time_t' or '__kernel_time_t' are part of
the user API that cannot be changed by that either have a
replacement or that do not suffer from the y2038 overflow.

Acked-by: Deepa Dinamani <deepa.kernel@gmail.com>
Acked-by: Christian Brauner <christian.brauner@ubuntu.com>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
(cherry picked from commit 2a785996cc5e2fc1d1d29d196f530905f68d2dc2)

Change-Id: Ifb7c359106bee111f6d524f6d65639141a3fdebc
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>