y2038: sparc: remove use of struct timex
'struct timex' is one of the last users of 'struct timeval' and is
only referenced in one place in the kernel any more, to convert the
user space timex into the kernel-internal version on sparc64, with a
different tv_usec member type.

As a preparation for hiding the time_t definition and everything
using that in the kernel, change the implementation once more
to only convert the timeval member, and then enclose the
struct definition in an #ifdef.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
(cherry picked from commit 251ec1c159e4874fbede0c3c586e317e177c0c9b)

Change-Id: Ia2858d105921efd438418418ad6f7f9aa54a6c9f
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>