quota: avoid time_t in v1_disk_dqblk definition
The time_t type is part of the user interface and not always the
same, with the move to 64-bit timestamps and the difference between
architectures.

Make the quota format definition independent of this type and use
a basic type of the same length. Make it unsigned in the process
to keep the v1 format working until year 2106 instead of 2038
on 32-bit architectures.

Hopefully, everybody has already moved to a newer format long
ago (v2 was introduced with linux-2.4), but it's hard to be sure.

Link: https://lore.kernel.org/r/20191213205221.3787308-6-arnd@arndb.de
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Jan Kara <jack@suse.cz>
(cherry picked from commit 1ead083ae1472e5aaee7f18ea827b13f55bd10a6)

Change-Id: I03f3e6158602a941ce1dc6378e131bba514bf908
Signed-off-by: Ronghua Zeng <quic_ronghuaz@quicinc.com>