skbuff_recycler: Limit max size of skb recycler
This change limits the maximum size of skbs that the
recycler can consume to 2688.

This change is added so that only the skbs allocated by
the ehternet and wifi driver from the skb_data_cache are allowed
into the recycler.

This is a temporary WAR to fix the memory leak in ezmesh scenario.

Change-Id: I53d09b8918152b75b1d0aefe9979583fd9ff3836
Signed-off-by: Aniruddha Bhat Anemajalu <quic_aanemaja@quicinc.com>