driver: bootconfig: Add CRC32 support in Bootconfig driver
CRC feature can be enabled while generating bootconfig binaries.
If this feature is enabled, magic end field will contain CRC value
instead of magic number.

This patch calculates CRC for bootconfig contents and
compares it with magic end. If one of the partitions BC0/BC1 is
corrupted, the contents are restored from SMEM.
If both the partitions are corrupted, proc entries are not created
so that sysupgrade will fail.

Change-Id: Ic001c95ac085aaa86b48fd429a60b8e4195eabf1
Signed-off-by: Gurumoorthy Santhakumar <quic_gsanthak@quicinc.com>