crypto: qce: Add command descriptor support for AEAD and SHA
Current scenario only supports command descriptor in Skciper.
It works in following steps:
1. Skciper will get a lock to use DMA
2. It will process with encryption/decryption
3. Unlock DMA for other processes

Since command descriptor is not enable in AEAD and SHA, it creates
a conflict with locking mechanism causing a crash as self test for
AEAD is failing further making crypto bad.

To overcome above issue add command descriptor support for AEAD
and update locking mechanism in AEAD and SHA to handle DMA.
This is to avoid crash due to timing issues in crypto.

Change-Id: I71d6b918bb422b2e6efb4a17daecf57bc9c5ba0e
Signed-off-by: Omeshwari wazekar <quic_owazekar@quicinc.com>