arm64: dts: ipq9574: Increase the MLO shared memory for RDP456
By default, the RDP456 has two groups namely group0 and group1.
Each group has two chips in it.

There is a rework to form single group0 and 4 MLO chips in it.
The WSI and MLO config is similar to RDP467.
Increase the MLO shared memory to 22MB to support 4 chips in a
group.
Without INI file, it will fallback to two group configurations.

Change-Id: I740b60c061c9b90341ba414706fa5b45e5b32638
Signed-off-by: Madhavan Ganesan <quic_madhgane@quicinc.com>