cfg80211: Add API to update current bss of non-assoc link vdev
wdev->current_bss is not updated for a non-assoc link STA vdevs.
As a result, a WARN_ON error is seen in cfg80211_ch_switch_notify()
when it’s called for non-assoc link.

To resolve this, update the wdev->current_bss for non-assoc links
as part of connection and clean up this as part of disconnection.

Change-Id: I7d739fc2420f2e66cd66e7ca6089b2921161f358
Signed-off-by: Shashikala Prabhu <quic_pshashik@quicinc.com>
CRs-Fixed: 3603593