drivers: qcom: bt_ipc: Cleanup deinit path only when the workqueues are actually initialized
When BTSS is enabled in uboot, it patches qcom,bt-running=1.
With this the BT rproc driver moves the state to RUNNING andskips bt_ipc_init
and WQs are not created.
At this stage, when rproc is stopped from sysfs, deinit API tries to destroy
non existent WQs.
Added a check in deinit to fix the same.

Change-Id: I79a26758e12ec1e01ccb0779ce92c17a5f14c25a
Signed-off-by: Poovendhan Selvaraj <quic_poovendh@quicinc.com>