skb-recycler: Add support to track the original caller
With MEMTRACER, SKBs allocated from recycler won't show the current
allocator of the SKB, rather it shows the allocator that initially
allocated it from the slab. This information is misleading.
To fix this, update the allocator of the SKB to the current call stack in
the following places.
  - skb_recycler_consume() - to reflect that the previous allocator has freed
                             it and now the recycler is the owning it
  - skb_recycler_alloc()   - to reflect that a new allocator is allocating
                             that SKB from recycler

Change-Id: I126b87cbbc7fc73a2e7533d595ba5bd2016e8bda
Signed-off-by: Shiva Krishna Pittala <quic_spittala@quicinc.com>
(cherry picked from commit b062358cc262c7d37e98416e66167a10a7290735)