soc: qcom: lm: Add ioctl support to get Bindings
Add GET_BINDINGS ioctl support to get the bindings from userspace.
Userspace needs to pass the 34 Bytes of NONCE and 2KB buffer for
bindings. LM driver will get the bindings from TMEL and copy it
to the ECDSA buffer in the response.

Userspace needs to pass the proper sizes in nonce_buf_len and
ecdsa_buf_len.

Change-Id: If58e56873742206fa17f6405435cc58e8afbefd5
Signed-off-by: Praveenkumar I <quic_ipkumar@quicinc.com>