drivers: remoteproc: qcom_q6v5: Complete pending waits on rproc crash
Rproc start/stop/spawn might be waiting for completion from IRQ when a
rproc crashes. In such cases, the recovery waits for the timeout before
proceeding as mutex is held by the boot/shutdown threads.

Complete all pending waits for an rproc when fatal or wdog IRQ is
received. In the wait_for_completion function, check if the
corresponding ack flag is set and return success/failure. This will
be set only if completion is from the IRQ and not from crash.

Also, for Multi-pd driver, complete pending waits for all userpds when
rootpd crashes.

Change-Id: Ia035799e83408800f0352a16cbdbf9e016bde6fe
Signed-off-by: Vignesh Viswanathan <quic_viswanat@quicinc.com>