drivers: pwm: ipq: pwm config sequence update
To configure the pwm wave, we need to follow
the below pwm_config sequence.

pwm_disable -> configure and apply divider values
-> pwm_enable

Instead of configuring the divider values in the
middle of the sequence, we can calculate it early
in order to shorten the duration for which the pwm
channel will disabled.

Change-Id: Id65b213e4cf8debb6f19afda957b59a2a0dea237
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>
(cherry picked from commit 2676445e0d4fb294560548ffa5c793e00cf70b41)