net: qrtr: Release the flows after a DEL_PROC
After a DEL_PROC is received, the waiters on pending flows are woken up.
But since the flows are not released, there is a chance that, the same
flow if its re-used, can have a stale pending count, that can get
incremented and finally result in a hang when HIGH watermark is reached.
So for every DEL_PROC release old flows.

Change-Id: I5335d89ecb9516c808f9a322b22f56a6f1b8305f
Signed-off-by: Sricharan R <sricharan@codeaurora.org>
Signed-off-by: Gokul Sriram Palanisamy <quic_gokulsri@quicinc.com>