drivers: pci: controller: dwc: qcom: Fix pci rcremove issue
With the wake gpio interrupt registered, while doing
pci rcremove, wake interrupt is received immediately due to
mhi reset and this causes the wake irq handler to rescan the
pci bus because of the low state of mdm2ap gpio.
Hence result in pci enumeration. Fixing this issue by checking
if mdm2ap interrupt is present before checking the gpio.

This also fixes a regression issue caused because of the wake gpio
interrupt support. pci crash would be seen during rcremove which was
caused because of the pcie->enumerated set to true before doing host
init. In cases where pci slot is enabled, but the link is not up the
pci->enumerated field will be set to true. This will cause crash
while doing rcremove as the pci device is itself not found.

Change-Id: Ib5192013a08f75a62cb33be5102460c195aae694
Signed-off-by: Hariharan K <quic_harihk@quicinc.com>