cfg80211: Avoid extra hold as part of connect done
As part of connection process for a n link MLO
the hold_bss is happening twice for non assoc links
as part of connect_done and update_current_bss path.
Consquently as part of disconnection/wifi
down there is a mismatch for the non assoc links
while doing unhold as there is an addtional hold in the
connect done path which is not getting cleared during the
wifi down path or disconnection.

Hence in order to prevent this refrain from taking the hold
for partner links and take the hold for only the assoc link.
So that the discreapany for the count of hold and unhold is
not seen

Change-Id: If21147cb7bd3c0885494ee6576f8f13081487b3a
Signed-off-by: Hari <hag@qti.qualcomm.com>