arm64: dts: Enable WiFi LED on IPQ8074 in 5.4 kernel
HK09 variant of IPQ8074 follows GPIO based LED mechanism
and it has been working with 4.4 kernel. In 5.4 kernel, the
GPIO APIs that were used in 4.4 kernel do not hold good and
have been flagged as a kernel bug with call trace and hence the
feature had been disabled from the wifi script. Enable it
back in the script along with corresponding code changes in
wifi source to have the appropriate GPIO APIs used for WiFi
LED to work in 5.4 kernel for HK09.

Change-Id: I6fe6c14bf97a61ab5988302cb8b4362097b23918
Signed-off-by: Sajini R <quic_sajinir@quicinc.com>